more templatizing
