+++ CMakeLists.txt
@@ -121,7 +121,7 @@ endif()
 
 # This set of sources is compiled twice, once in single precision and once in
 # double precision The single precision compilation is done with -DSINGLE
-set(FINUFFT_PRECISION_DEPENDENT_SOURCES src/finufft.cpp src/fft.cpp
+set(FINUFFT_PRECISION_DEPENDENT_SOURCES src/finufft.cpp
                                         src/simpleinterfaces.cpp)
 
 # If we're building for Fortran, make sure we also include the translation
@@ -258,11 +258,15 @@ if(FINUFFT_USE_CPU)
   add_library(finufft_f64 OBJECT ${FINUFFT_PRECISION_DEPENDENT_SOURCES})
   set_finufft_options(finufft_f64)
   if(NOT FINUFFT_STATIC_LINKING)
-    add_library(finufft SHARED src/spreadinterp.cpp src/utils_precindep.cpp
-                               contrib/legendre_rule_fast.cpp)
+    add_library(
+      finufft SHARED
+      src/spreadinterp.cpp src/utils_precindep.cpp
+      contrib/legendre_rule_fast.cpp src/fft.cpp src/finufft_core.cpp)
   else()
-    add_library(finufft STATIC src/spreadinterp.cpp src/utils_precindep.cpp
-                               contrib/legendre_rule_fast.cpp)
+    add_library(
+      finufft STATIC
+      src/spreadinterp.cpp src/utils_precindep.cpp
+      contrib/legendre_rule_fast.cpp src/fft.cpp src/finufft_core.cpp)
   endif()
   target_link_libraries(finufft PRIVATE finufft_f32 finufft_f64)
   set_finufft_options(finufft)
