+++ include/finufft/utils_precindep.h
@@ -4,9 +4,9 @@
 #ifndef UTILS_PRECINDEP_H
 #define UTILS_PRECINDEP_H
 
-#include "defs.h"
-// for CNTime...
-// using chrono since the interface is portable between linux and windows
+// #include "defs.h"
+//  for CNTime...
+//  using chrono since the interface is portable between linux and windows
 #include <chrono>
 
 namespace finufft {
