+++ test/testutils.cpp
@@ -57,7 +57,8 @@ int main(int argc, char *argv[]) {
     a[j] = CPX(1.0, 0.0);
     b[j] = a[j];
   }
-  FLT relerr = 2.0 * EPSILON; // 1 ULP, fine since 1.0 rep exactly
+  constexpr FLT EPSILON = std::numeric_limits<FLT>::epsilon();
+  FLT relerr            = 2.0 * EPSILON; // 1 ULP, fine since 1.0 rep exactly
   if (abs(infnorm(M, &a[0]) - 1.0) > relerr) return 1;
   if (abs(twonorm(M, &a[0]) - sqrt((FLT)M)) > relerr * sqrt((FLT)M)) return 1;
   b[0] = CPX(0.0, 0.0); // perturb b from a
